:truck: rename: 도메인 디렉토리 구조로 변경 #9
