--- src/main/java/com/habitpay/habitpay/config/auth/dto/SessionUser.java
@@ -1,17 +1,0 @@
-package com.habitpay.habitpay.config.auth.dto;
-
-import com.habitpay.habitpay.user.entities.User;
-import lombok.Getter;
-
-import java.io.Serializable;
-
-@Getter
-public class SessionUser implements Serializable {
-    private String name;
-    private String email;
-
-    public SessionUser(User user) {
-        this.name = user.getName();
-        this.email = user.getEmail();
-    }
-}
