+++ src/main/java/com/habitpay/habitpay/domain/member/dao/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.habitpay.habitpay.domain.member.dao;
+
+import com.habitpay.habitpay.domain.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String email);
+}
