+++ src/main/java/com/habitpay/habitpay/domain/member/domain/Member.java
@@ -1,6 +1,6 @@
-package com.habitpay.habitpay.user.entities;
+package com.habitpay.habitpay.domain.member.domain;
 
-import com.habitpay.habitpay.user.enums.Role;
+import com.habitpay.habitpay.domain.model.BaseTime;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,8 +9,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @Entity
-@Table(name = "users")
-public class User extends BaseTimeEntity {
+@Table(name = "members")
+public class Member extends BaseTime {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -26,13 +26,13 @@ public class User extends BaseTimeEntity {
     private Role role;
 
     @Builder
-    public User(String name, String email, Role role) {
+    public Member(String name, String email, Role role) {
         this.name = name;
         this.email = email;
         this.role = role;
     }
 
-    public User update(String name) {
+    public Member update(String name) {
         this.name = name;
 
         return this;
