+++ src/main/java/com/habitpay/habitpay/global/config/auth/SecurityConfig.java
@@ -1,13 +1,12 @@
-package com.habitpay.habitpay.config.auth;
+package com.habitpay.habitpay.global.config.auth;
 
-import com.habitpay.habitpay.user.enums.Role;
+import com.habitpay.habitpay.domain.member.domain.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
