+++ src/main/java/com/habitpay/habitpay/global/config/auth/dto/SessionUser.java
@@ -1,0 +1,17 @@
+package com.habitpay.habitpay.global.config.auth.dto;
+
+import com.habitpay.habitpay.domain.member.domain.Member;
+import lombok.Getter;
+
+import java.io.Serializable;
+
+@Getter
+public class SessionUser implements Serializable {
+    private String name;
+    private String email;
+
+    public SessionUser(Member member) {
+        this.name = member.getName();
+        this.email = member.getEmail();
+    }
+}
