--- src/main/java/com/habitpay/habitpay/user/repositories/UserRepository.java
@@ -1,10 +1,0 @@
-package com.habitpay.habitpay.user.repositories;
-
-import com.habitpay.habitpay.user.entities.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-}
